Refactor of user settings, add new PassAutcomplete
- index.tsx used to literally only wrap PasswordForm in a <Pane>, it is
now all in index.tsx
- Made it possible to give any HTMLInput-prop to PasswordInput
- Gave the new password fields autoComplete="new-password", so
passwordmanagers can auto complete it easier
